Wrapped all renderPlots inside observeEvent calls
Added a bunch of observeEvent() calls in server script to make sure
the input variables (menu items) have been selected before the server
tries to render any of the plots. This is mostly a cosmetic change
because it gets rid of some annoying warnings and glitchy stuff that
used to occur when the app first loads.