Example of how to make scripts standalone
app.R sets a global variable called "INAPP". A few lines of code at the
top of a plotting script checks if this global "INAPP" variable exists
and knows if it is being sourced standalone (for troubleshooting) or if
it is being sourced by the shiny app.

This saves the trouble of commenting/uncommenting lines of code at
the top of each plotting script when troubleshooting.